Move objects update function back to object.js file from index
