+++ src/objects.test.js
@@ -1,9 +1,11 @@
 import { 
   createGameObject,
+  Direction,
   getObjectsIntersectionType,
   getObjectsIntersectionRect,
   ObjectsIntersectionType,
-} from './objects';
+  updateObject,
+} from "./objects";
 
 
 describe("Objects intersection", () => {
@@ -133,7 +135,7 @@ describe("Objects intersection", () => {
     });
   });
 
-  describe("ObjectsIntersectionType", () => {
+  describe("Objects Intersection Type", () => {
     test("Objects don't intersect", () => {
       const randomObject1 = createGameObject(10, 10, 10, 10);
       const randomObject2 = createGameObject(100, 100, 10, 10);
@@ -197,3 +199,36 @@ describe("Objects intersection", () => {
     });
   });
 });
+
+describe("Movement tests", () => {
+  describe("Gravity tests", () => {
+    test("Object moves down with its vertical speed when it's not staying on the ground", () => {
+      const initialObject = createGameObject(
+        0, 10, 10, 10, 
+        Direction.RIGHT | Direction.DOWN,
+        0, 5
+      );
+      expect(updateObject(initialObject).y).toEqual(5);
+    });
+
+    test("Object doesn't move down if it stays on the ground (y = 0)", () => {
+      const initialObject = createGameObject(
+        0, 0, 10, 10, 
+        Direction.RIGHT | Direction.DOWN,
+        0, 5
+      );
+      expect(updateObject(initialObject).y).toEqual(0);
+    });
+
+    test("It is impossible for the object to fall below ground: negative Y coord is turned to 0", () => {
+      const initialObject = createGameObject(
+        0, -5, 10, 10, 
+        Direction.RIGHT | Direction.DOWN,
+        0, 5
+      );
+
+      expect(updateObject(initialObject).y).toEqual(0);
+    });
+  });
+});
+
